refactor: naming and code improvements
