+++ src/ui/page/dataverse/dataspace/dataspace.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react'
 import { useParams } from 'react-router-dom'
 import type { Option } from 'fp-ts/Option'
 import { getResourceDetails } from '@/ui/page/dataverse/dataverse'
-import { NotFoundError } from '../../error/notFound/notFoundError'
+import { NotFoundError } from '../../notFoundError/notFoundError'
 import type { DataSpace, DataverseItemDetails } from '@/ui/page/dataverse/dataverse'
 import type { ItemGeneralMetadata } from '@/ui/view/dataverse/types'
 import PageTemplate from '@/ui/view/dataverse/component/pageTemplate/pageTemplate'
