+++ src/ui/page/dataverse/service/service.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react'
 import { useParams } from 'react-router-dom'
 import type { Option } from 'fp-ts/Option'
 import { match, none } from 'fp-ts/Option'
-import { NotFoundError } from '../../error/notFound/notFoundError'
+import { NotFoundError } from '../../notFoundError/notFoundError'
 import { getResourceDetails } from '@/ui/page/dataverse/dataverse'
 import type { DataverseItemDetails } from '@/ui/page/dataverse/dataverse'
 import type { ItemGeneralMetadata } from '@/ui/view/dataverse/types'
