--- src/ui/page/error/i18n/error_de.json
@@ -1,11 +1,0 @@
-{
-  "error": {
-    "notFoundError": {
-      "apology": "Entschuldigung,",
-      "message": "wir können die Seite, die Sie suchen, nicht finden.",
-      "suggestion": "Versuchen Sie, die URL zu überprüfen, oder vielleicht möchten Sie:",
-      "dataverseButton": "Das Datenverse erkunden",
-      "homeButton": "Zurück zur Startseite"
-    }
-  }
-}
