--- src/ui/page/error/i18n/error_en.json
@@ -1,11 +1,0 @@
-{
-  "error": {
-    "notFoundError": {
-      "apology": "Sorry,",
-      "message": "We can't find the page you are looking for.",
-      "suggestion": "Try checking the URL or maybe you want to:",
-      "dataverseButton": "Explore the dataverse",
-      "homeButton": "Go to the homepage"
-    }
-  }
-}
