--- src/ui/page/error/i18n/error_fr.json
@@ -1,11 +1,0 @@
-{
-  "error": {
-    "notFoundError": {
-      "apology": "Désolé,",
-      "message": "Nous ne trouvons pas la page que vous avez demandée .",
-      "suggestion": "Verifiez l'URL ou vous pouvez :",
-      "dataverseButton": "Explorer le dataverse",
-      "homeButton": "Retourner à l'accueil"
-    }
-  }
-}
