--- src/ui/page/error/notFound/notFoundError.scss
@@ -1,90 +1,0 @@
-@import '@/ui/style/fonts.scss';
-@import '@/ui/style/theme.scss';
-@import '@/ui/style/mixins.scss';
-
-// .okp4-dataverse-portal-page-layout {
-.okp4-dataverse-portal-not-found-error-page-main {
-  @include with-theme() {
-    &.dark {
-      background: url('../../../asset/images/error-not-found-gradient-dark.webp'),
-        url('../../../asset/images/error-not-found-shadow-dark.webp'),
-        url('../../../asset/images/error-not-found-wood-texture-dark.webp');
-      background-size: cover;
-    }
-
-    &.light {
-      background: url('../../../asset/images/error-not-found-gradient-light.webp'),
-        url('../../../asset/images/error-not-found-shadow-light.webp'),
-        url('../../../asset/images/error-not-found-wood-texture-light.webp');
-      background-size: cover;
-    }
-
-    background-blend-mode: overlay;
-    height: 100%;
-
-    .okp4-dataverse-portal-not-found-error-page-content-wrapper {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      justify-content: center;
-      text-align: center;
-      gap: 28px;
-      height: 100%;
-
-      @include use-breakpoints(('mobile', 'tablet')) {
-        gap: 24px;
-      }
-
-      h1 {
-        @include norse-font(700);
-        letter-spacing: 0.0075em;
-        line-height: 88px;
-        font-size: 80px;
-        color: themed('text');
-
-        @include use-breakpoints(('mobile', 'tablet')) {
-          font-size: 40px;
-          line-height: 48px;
-        }
-      }
-
-      .okp4-dataverse-portal-not-found-error-page-message-titles {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        max-width: 750px;
-
-        .okp4-dataverse-portal-not-found-error-page-message-apology {
-          opacity: 0.4;
-        }
-      }
-
-      .okp4-dataverse-portal-not-found-error-page-message-suggestion {
-        @include noto-sans(300);
-        font-size: 18px;
-        color: themed('text');
-        line-height: 22px;
-      }
-
-      .okp4-dataverse-portal-not-found-error-page-button-container {
-        display: flex;
-        gap: 20px;
-
-        @include use-breakpoints(('mobile')) {
-          flex-direction: column;
-          width: 100%;
-        }
-
-        .okp4-dataverse-portal-not-found-error-page-button {
-          max-height: 40px;
-          width: fit-content;
-
-          @include use-breakpoints(('mobile')) {
-            width: 100%;
-          }
-        }
-      }
-    }
-  }
-}
-// }
