+++ src/ui/page/notFoundError/i18n/notFoundError_de.json
@@ -1,0 +1,9 @@
+{
+  "notFoundError": {
+    "apology": "Entschuldigung,",
+    "message": "wir können die Seite, die Sie suchen, nicht finden.",
+    "suggestion": "Versuchen Sie, die URL zu überprüfen, oder vielleicht möchten Sie:",
+    "backToDataverse": "Das Datenverse erkunden",
+    "backToHome": "Zurück zur Startseite"
+  }
+}
