+++ src/ui/page/notFoundError/i18n/notFoundError_fr.json
@@ -1,0 +1,9 @@
+{
+  "notFoundError": {
+    "apology": "Désolé,",
+    "message": "Nous ne trouvons pas la page que vous avez demandée .",
+    "suggestion": "Verifiez l'URL ou vous pouvez :",
+    "backToDataverse": "Explorer le dataverse",
+    "backToHome": "Retourner à l'accueil"
+  }
+}
